+++ src/heuristics.jl
@@ -706,27 +706,30 @@ function refinement_tree_1opt!(
     adjacency_augment_graph::Matrix{<:Number},
     edge_to_check::Graphs.SimpleGraphs.SimpleEdge{<:Number},
 )
-    if !(Graphs.has_edge(G, Graphs.src(edge_to_check), Graphs.dst(edge_to_check)))
-        Graphs.add_edge!(G, Graphs.src(edge_to_check), Graphs.dst(edge_to_check))
+    src, dst = Graphs.src(edge_to_check), Graphs.dst(edge_to_check)
+    if !Graphs.has_edge(G, src, dst)
+        Graphs.add_edge!(G, src, dst)
         ac_tracker = 0.0
-        vertices_tracker = (undef, undef)
+        best_edge = nothing
+        
+        combined_adjacency = adjacency_augment_graph + adjacency_base_graph
+        
         for edge in Graphs.edges(G)
-            if adjacency_base_graph[Graphs.src(edge), Graphs.dst(edge)] == 0
-                Graphs.rem_edge!(G, Graphs.src(edge), Graphs.dst(edge))
-                if LOpt.algebraic_connectivity(
-                    (adjacency_augment_graph + adjacency_base_graph) .*
-                    Matrix(Graphs.adjacency_matrix(G)),
-                ) > ac_tracker
-                    ac_tracker = LOpt.algebraic_connectivity(
-                        (adjacency_augment_graph + adjacency_base_graph) .*
-                        Matrix(Graphs.adjacency_matrix(G)),
-                    )
-                    vertices_tracker = edge
+            s, d = Graphs.src(edge), Graphs.dst(edge)
+            if adjacency_base_graph[s, d] == 0
+                Graphs.rem_edge!(G, s, d)
+                ac = LOpt.algebraic_connectivity(combined_adjacency .* Matrix(Graphs.adjacency_matrix(G)))
+                if ac > ac_tracker
+                    ac_tracker = ac
+                    best_edge = edge
                 end
-                Graphs.add_edge!(G, Graphs.src(edge), Graphs.dst(edge))
+                Graphs.add_edge!(G, s, d)
             end
         end
-        Graphs.rem_edge!(G, vertices_tracker)
+        
+        if best_edge !== nothing
+            Graphs.rem_edge!(G, best_edge)
+        end
     end
     return G
 end
@@ -738,18 +741,29 @@ function refinement_tree_2opt!(
     edge_to_check_1::Graphs.SimpleGraphs.SimpleEdge{<:Number},
     edge_to_check_2::Graphs.SimpleGraphs.SimpleEdge{<:Number},
 )
-    if !(Graphs.has_edge(G, Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1))) &&
-       !(Graphs.has_edge(G, Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2)))
-        G = LOpt.add_multiple_edges!(
-            G,
-            [
-                (Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1)),
-                (Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2)),
-            ],
-        )
+    # Check which edges need to be added
+    edge1_exists = Graphs.has_edge(G, Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1))
+    edge2_exists = Graphs.has_edge(G, Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2))
+    
+    if edge1_exists && edge2_exists
+        return G
+    # One edge is missing, use 1-opt refinement
+    elseif edge1_exists
+        return LOpt.refinement_tree_1opt!(G, adjacency_base_graph, adjacency_augment_graph, edge_to_check_2)
+    elseif edge2_exists
+        return LOpt.refinement_tree_1opt!(G, adjacency_base_graph, adjacency_augment_graph, edge_to_check_1)
+    # Both edges are missing, add them and find optimal edges to remove
+    else
+        edges_to_add = [
+            (Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1)),
+            (Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2))
+        ]
+        G = LOpt.add_multiple_edges!(G, edges_to_add)
+        
         ac_tracker = 0.0
         vertices_tracker = [(undef, undef), (undef, undef)]
-        ac_tracker, vertices_tracker = _vertices_tracker_update_two_edges!(
+        
+        ac_tracker, vertices_tracker = LOpt._vertices_tracker_update_two_edges!(
             G,
             adjacency_base_graph,
             adjacency_augment_graph,
@@ -757,31 +771,9 @@ function refinement_tree_2opt!(
             ac_tracker,
             vertices_tracker,
         )
-        G = LOpt.remove_multiple_edges!(G, vertices_tracker)
-    elseif !(Graphs.has_edge(
-        G,
-        Graphs.src(edge_to_check_1),
-        Graphs.dst(edge_to_check_1),
-    )) && (Graphs.has_edge(G, Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2)))
-        G = refinement_tree_1opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_1,
-        )
-    elseif !(Graphs.has_edge(
-        G,
-        Graphs.src(edge_to_check_2),
-        Graphs.dst(edge_to_check_2),
-    )) && (Graphs.has_edge(G, Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1)))
-        G = refinement_tree_1opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_2,
-        )
+        
+        return LOpt.remove_multiple_edges!(G, vertices_tracker)
     end
-    return G
 end
 
 function refinement_tree_3opt!(
@@ -792,27 +784,49 @@ function refinement_tree_3opt!(
     edge_to_check_2::Graphs.SimpleGraphs.SimpleEdge{<:Number},
     edge_to_check_3::Graphs.SimpleGraphs.SimpleEdge{<:Number},
 )
-    con1 = !(Graphs.has_edge(G, Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1)))
-    con2 = !(Graphs.has_edge(G, Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2)))
-    con3 = !(Graphs.has_edge(G, Graphs.src(edge_to_check_3), Graphs.dst(edge_to_check_3)))
-    if con1 && con2 && con3
-        G = LOpt.add_multiple_edges!(
+    # Check which edges are not in the graph
+    edges_to_add = []
+    for (i, edge) in enumerate([edge_to_check_1, edge_to_check_2, edge_to_check_3])
+        if !Graphs.has_edge(G, Graphs.src(edge), Graphs.dst(edge))
+            push!(edges_to_add, (edge, i))
+        end
+    end
+    
+    num_edges = length(edges_to_add)
+    
+    if num_edges == 0
+        return G
+    elseif num_edges == 1
+        edge = edges_to_add[1][1]
+        return refinement_tree_1opt!(
             G,
-            [
-                (Graphs.src(edge_to_check_1), Graphs.dst(edge_to_check_1)),
-                (Graphs.src(edge_to_check_2), Graphs.dst(edge_to_check_2)),
-                (Graphs.src(edge_to_check_3), Graphs.dst(edge_to_check_3)),
-            ],
+            adjacency_base_graph,
+            adjacency_augment_graph,
+            edge,
         )
+    elseif num_edges == 2
+        edge1 = edges_to_add[1][1]
+        edge2 = edges_to_add[2][1]
+        return refinement_tree_2opt!(
+            G,
+            adjacency_base_graph,
+            adjacency_augment_graph,
+            edge1,
+            edge2,
+        )
+    else # num_edges == 3
+        # Add all three edges
+        edge_pairs = [(Graphs.src(e[1]), Graphs.dst(e[1])) for e in edges_to_add]
+        G = LOpt.add_multiple_edges!(G, edge_pairs)
+        
+        # Track best configuration
         ac_tracker = 0.0
         vertices_tracker = [(undef, undef), (undef, undef), (undef, undef)]
+        
+        # Check edges that can be removed
         for j in 1:(length(collect(Graphs.edges(G)))-2)
-            if (
-                adjacency_base_graph[
-                    Graphs.src(collect(Graphs.edges(G))[j]),
-                    Graphs.dst(collect(Graphs.edges(G))[j]),
-                ] == 0
-            )
+            edge = collect(Graphs.edges(G))[j]
+            if adjacency_base_graph[Graphs.src(edge), Graphs.dst(edge)] == 0
                 ac_tracker, vertices_tracker = LOpt._vertices_tracker_update_two_edges!(
                     G,
                     adjacency_base_graph,
@@ -823,52 +837,7 @@ function refinement_tree_3opt!(
                 )
             end
         end
-        G = LOpt.remove_multiple_edges!(G, vertices_tracker)
-    elseif con1 && con2 && !(con3)
-        G = refinement_tree_2opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_1,
-            edge_to_check_2,
-        )
-    elseif con1 && !(con2) && con3
-        G = refinement_tree_2opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_1,
-            edge_to_check_3,
-        )
-    elseif !(con1) && con2 && con3
-        G = refinement_tree_2opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_2,
-            edge_to_check_3,
-        )
-    elseif con1 && !(con2) && !(con3)
-        G = refinement_tree_1opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_1,
-        )
-    elseif !(con1) && con2 && !(con3)
-        G = refinement_tree_1opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_2,
-        )
-    elseif !(con1) && !(con2) && con3
-        G = refinement_tree_1opt!(
-            G,
-            adjacency_base_graph,
-            adjacency_augment_graph,
-            edge_to_check_3,
-        )
+        
+        return LOpt.remove_multiple_edges!(G, vertices_tracker)
     end
-    return G
 end